// plugwerk-server/plugwerk-server-frontend/src/api/generated/model/plugin-release-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plugwerk API
 * Plugin marketplace for the Java/PF4J ecosystem
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PluginDependencyDto } from './plugin-dependency-dto';

/**
 * 
 * @export
 * @interface PluginReleaseDto
 */
export interface PluginReleaseDto {
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'pluginId': string;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'version': string;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'changelog'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'artifactSha256'?: string;
    /**
     * 
     * @type {number}
     * @memberof PluginReleaseDto
     */
    'artifactSize'?: number;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'requiresSystemVersion'?: string;
    /**
     * 
     * @type {number}
     * @memberof PluginReleaseDto
     */
    'requiresApiLevel'?: number;
    /**
     * 
     * @type {Array<PluginDependencyDto>}
     * @memberof PluginReleaseDto
     */
    'pluginDependencies'?: Array<PluginDependencyDto>;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'status': PluginReleaseDtoStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'publishedAt'?: string;
    /**
     * 
     * @type {number}
     * @memberof PluginReleaseDto
     */
    'downloadCount'?: number;
    /**
     * 
     * @type {string}
     * @memberof PluginReleaseDto
     */
    'createdAt'?: string;
}

export const PluginReleaseDtoStatusEnum = {
    Draft: 'draft',
    Published: 'published',
    Deprecated: 'deprecated',
    Yanked: 'yanked'
} as const;

export type PluginReleaseDtoStatusEnum = typeof PluginReleaseDtoStatusEnum[keyof typeof PluginReleaseDtoStatusEnum];